MEP 11/09/2024

+++ src/publicodes-state/providers/userProvider/useUpdateOldLocalStorage.ts
@@ -4,7 +4,6 @@ import {
   MigrationType,
   Simulation,
 } from '@/publicodes-state/types'
-import { captureException } from '@sentry/react'
 import { useEffect } from 'react'
 
 type Props = {
@@ -18,21 +17,22 @@ function handleLocalStorageMigration(
   migrationInstructions: MigrationType
 ) {
   try {
-    const newSimulations = currentLocalStorage.simulations.map(
-      (simulation: Simulation) =>
-        migrateSimulation({
-          simulation,
-          migrationInstructions,
-        })
-    )
+    if(currentLocalStorage.simulations) {
+      const newSimulations = currentLocalStorage.simulations.map(
+          (simulation: Simulation) =>
+              migrateSimulation({
+                simulation,
+                migrationInstructions,
+              })
+      )
 
-    localStorage.setItem(
-      storageKey,
-      JSON.stringify({ ...currentLocalStorage, simulations: newSimulations })
-    )
+      localStorage.setItem(
+          storageKey,
+          JSON.stringify({ ...currentLocalStorage, simulations: newSimulations })
+      )
+    }
   } catch (error) {
     console.warn('Error trying to migrate LocalStorage:', error)
-    captureException(error)
   }
 }
